// src/types/anilistv2.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  CountryCode: any;
  FuzzyDateInt: any;
  Json: any;
};
// ...
export type Favourites = {
  /** Favourite anime */
  anime?: Maybe<MediaConnection>;
  /** Favourite characters */
  characters?: Maybe<CharacterConnection>;
  /** Favourite manga */
  manga?: Maybe<MediaConnection>;
  /** Favourite staff */
  staff?: Maybe<StaffConnection>;
  /** Favourite studios */
  studios?: Maybe<StudioConnection>;
};
// ...
export type MediaListOptions = {
  /** The user's anime list options */
  animeList?: Maybe<MediaListTypeOptions>;
  /** The user's manga list options */
  mangaList?: Maybe<MediaListTypeOptions>;
  /** The default order list rows should be displayed in */
  rowOrder?: Maybe<Scalars["String"]>;
  /** The score format the user is using for media lists */
  scoreFormat?: Maybe<ScoreFormat>;
  /**
   * The list theme options for both lists
   * @deprecated No longer used
   */
  sharedTheme?: Maybe<Scalars["Json"]>;
  /**
   * If the shared theme should be used instead of the individual list themes
   * @deprecated No longer used
   */
  sharedThemeEnabled?: Maybe<Scalars["Boolean"]>;
  /** @deprecated No longer used */
  useLegacyLists?: Maybe<Scalars["Boolean"]>;
};
// ...
export enum MediaType {
  /** Japanese Anime */
  Anime = "ANIME",
  /** Asian comic */
  Manga = "MANGA",
}
// ...
export enum ModRole {
  /** An AniList administrator */
  Admin = "ADMIN",
  /** An anime data moderator */
  AnimeData = "ANIME_DATA",
  /** A community moderator */
  Community = "COMMUNITY",
  /** An AniList developer */
  Developer = "DEVELOPER",
  /** A discord community moderator */
  DiscordCommunity = "DISCORD_COMMUNITY",
  /** A lead anime data moderator */
  LeadAnimeData = "LEAD_ANIME_DATA",
  /** A lead community moderator */
  LeadCommunity = "LEAD_COMMUNITY",
  /** A head developer of AniList */
  LeadDeveloper = "LEAD_DEVELOPER",
  /** A lead manga data moderator */
  LeadMangaData = "LEAD_MANGA_DATA",
  /** A lead social media moderator */
  LeadSocialMedia = "LEAD_SOCIAL_MEDIA",
  /** A manga data moderator */
  MangaData = "MANGA_DATA",
  /** A retired moderator */
  Retired = "RETIRED",
  /** A social media moderator */
  SocialMedia = "SOCIAL_MEDIA",
}
// ...
export type Review = {
  /** The main review body text */
  body?: Maybe<Scalars["String"]>;
  /** The time of the thread creation */
  createdAt: Scalars["Int"];
  /** The id of the review */
  id: Scalars["Int"];
  /** The media the review is of */
  media?: Maybe<Media>;
  /** The id of the review's media */
  mediaId: Scalars["Int"];
  /** For which type of media the review is for */
  mediaType?: Maybe<MediaType>;
  /** If the review is not yet publicly published and is only viewable by creator */
  private?: Maybe<Scalars["Boolean"]>;
  /** The total user rating of the review */
  rating?: Maybe<Scalars["Int"]>;
  /** The amount of user ratings of the review */
  ratingAmount?: Maybe<Scalars["Int"]>;
  /** The review score of the media */
  score?: Maybe<Scalars["Int"]>;
  /** The url for the review page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
  /** A short summary of the review */
  summary?: Maybe<Scalars["String"]>;
  /** The time of the thread last update */
  updatedAt: Scalars["Int"];
  /** The creator of the review */
  user?: Maybe<User>;
  /** The id of the review's creator */
  userId: Scalars["Int"];
  /** The rating of the review by currently authenticated user */
  userRating?: Maybe<ReviewRating>;
};
// ...
export enum ReviewRating {
  DownVote = "DOWN_VOTE",
  NoVote = "NO_VOTE",
  UpVote = "UP_VOTE",
}
// ...
export type Thread = {
  /** The text body of the thread (Markdown) */
  body?: Maybe<Scalars["String"]>;
  /** The categories of the thread */
  categories?: Maybe<Array<Maybe<ThreadCategory>>>;
  /** The time of the thread creation */
  createdAt: Scalars["Int"];
  /** The id of the thread */
  id: Scalars["Int"];
  /** If the currently authenticated user liked the thread */
  isLiked?: Maybe<Scalars["Boolean"]>;
  /** If the thread is locked and can receive comments */
  isLocked?: Maybe<Scalars["Boolean"]>;
  /** If the thread is stickied and should be displayed at the top of the page */
  isSticky?: Maybe<Scalars["Boolean"]>;
  /** If the currently authenticated user is subscribed to the thread */
  isSubscribed?: Maybe<Scalars["Boolean"]>;
  /** The amount of likes the thread has */
  likeCount: Scalars["Int"];
  /** The users who liked the thread */
  likes?: Maybe<Array<Maybe<User>>>;
  /** The media categories of the thread */
  mediaCategories?: Maybe<Array<Maybe<Media>>>;
  /** The time of the last reply */
  repliedAt?: Maybe<Scalars["Int"]>;
  /** The id of the most recent comment on the thread */
  replyCommentId?: Maybe<Scalars["Int"]>;
  /** The number of comments on the thread */
  replyCount?: Maybe<Scalars["Int"]>;
  /** The user to last reply to the thread */
  replyUser?: Maybe<User>;
  /** The id of the user who most recently commented on the thread */
  replyUserId?: Maybe<Scalars["Int"]>;
  /** The url for the thread page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
  /** The title of the thread */
  title?: Maybe<Scalars["String"]>;
  /** The time of the thread last update */
  updatedAt: Scalars["Int"];
  /** The owner of the thread */
  user?: Maybe<User>;
  /** The id of the thread owner user */
  userId: Scalars["Int"];
  /** The number of times users have viewed the thread */
  viewCount?: Maybe<Scalars["Int"]>;
};

/** Forum Thread */
export type ThreadBodyArgs = {
  asHtml?: InputMaybe<Scalars["Boolean"]>;
};

/** A forum thread category */
export type ThreadCategory = {
  /** The id of the category */
  id: Scalars["Int"];
  /** The name of the category */
  name: Scalars["String"];
};
// ...
export type User = {
  /** The bio written by user (Markdown) */
  about?: Maybe<Scalars["String"]>;
  /** The user's avatar images */
  avatar?: Maybe<UserAvatar>;
  /** The user's banner images */
  bannerImage?: Maybe<Scalars["String"]>;
  bans?: Maybe<Scalars["Json"]>;
  /** When the user's account was created. (Does not exist for accounts created before 2020) */
  createdAt?: Maybe<Scalars["Int"]>;
  /** Custom donation badge text */
  donatorBadge?: Maybe<Scalars["String"]>;
  /** The donation tier of the user */
  donatorTier?: Maybe<Scalars["Int"]>;
  /** The users favourites */
  favourites?: Maybe<Favourites>;
  /** The id of the user */
  id: Scalars["Int"];
  /** If the user is blocked by the authenticated user */
  isBlocked?: Maybe<Scalars["Boolean"]>;
  /** If this user if following the authenticated user */
  isFollower?: Maybe<Scalars["Boolean"]>;
  /** If the authenticated user if following this user */
  isFollowing?: Maybe<Scalars["Boolean"]>;
  /** The user's media list options */
  mediaListOptions?: Maybe<MediaListOptions>;
  /** The user's moderator roles if they are a site moderator */
  moderatorRoles?: Maybe<Array<Maybe<ModRole>>>;
  /**
   * If the user is a moderator or data moderator
   * @deprecated Deprecated. Replaced with moderatorRoles field.
   */
  moderatorStatus?: Maybe<Scalars["String"]>;
  /** The name of the user */
  name: Scalars["String"];
  /** The user's general options */
  options?: Maybe<UserOptions>;
  /** The user's previously used names. */
  previousNames?: Maybe<Array<Maybe<UserPreviousName>>>;
  /** The url for the user page on the AniList website */
  siteUrl?: Maybe<Scalars["String"]>;
  /** The users anime & manga list statistics */
  statistics?: Maybe<UserStatisticTypes>;
  /**
   * The user's statistics
   * @deprecated Deprecated. Replaced with statistics field.
   */
  stats?: Maybe<UserStats>;
  /** The number of unread notifications the user has */
  unreadNotificationCount?: Maybe<Scalars["Int"]>;
  /** When the user's data was last updated */
  updatedAt?: Maybe<Scalars["Int"]>;
};
// ...
export type UserAvatar = {
  /** The avatar of user at its largest size */
  large?: Maybe<Scalars["String"]>;
  /** The avatar of user at medium size */
  medium?: Maybe<Scalars["String"]>;
};
// ...
export type UserOptions = {
  /** Minutes between activity for them to be merged together. 0 is Never, Above 2 weeks (20160 mins) is Always. */
  activityMergeTime?: Maybe<Scalars["Int"]>;
  /** Whether the user receives notifications when a show they are watching aires */
  airingNotifications?: Maybe<Scalars["Boolean"]>;
  /** The list activity types the user has disabled from being created from list updates */
  disabledListActivity?: Maybe<Array<Maybe<ListActivityOption>>>;
  /** Whether the user has enabled viewing of 18+ content */
  displayAdultContent?: Maybe<Scalars["Boolean"]>;
  /** Notification options */
  notificationOptions?: Maybe<Array<Maybe<NotificationOption>>>;
  /** Profile highlight color (blue, purple, pink, orange, red, green, gray) */
  profileColor?: Maybe<Scalars["String"]>;
  /** Whether the user only allow messages from users they follow */
  restrictMessagesToFollowing?: Maybe<Scalars["Boolean"]>;
  /** The language the user wants to see staff and character names in */
  staffNameLanguage?: Maybe<UserStaffNameLanguage>;
  /** The user's timezone offset (Auth user only) */
  timezone?: Maybe<Scalars["String"]>;
  /** The language the user wants to see media titles in */
  titleLanguage?: Maybe<UserTitleLanguage>;
};

/** A user's previous name */
export type UserPreviousName = {
  /** When the user first changed from this name. */
  createdAt?: Maybe<Scalars["Int"]>;
  /** A previous name of the user. */
  name?: Maybe<Scalars["String"]>;
  /** When the user most recently changed from this name. */
  updatedAt?: Maybe<Scalars["Int"]>;
};
// ...
export type UserStatisticTypes = {
  anime?: Maybe<UserStatistics>;
  manga?: Maybe<UserStatistics>;
};
// ...
export type UserStats = {
  activityHistory?: Maybe<Array<Maybe<UserActivityHistory>>>;
  animeListScores?: Maybe<ListScoreStats>;
  animeScoreDistribution?: Maybe<Array<Maybe<ScoreDistribution>>>;
  animeStatusDistribution?: Maybe<Array<Maybe<StatusDistribution>>>;
  /** The amount of manga chapters the user has read */
  chaptersRead?: Maybe<Scalars["Int"]>;
  favouredActors?: Maybe<Array<Maybe<StaffStats>>>;
  favouredFormats?: Maybe<Array<Maybe<FormatStats>>>;
  favouredGenres?: Maybe<Array<Maybe<GenreStats>>>;
  favouredGenresOverview?: Maybe<Array<Maybe<GenreStats>>>;
  favouredStaff?: Maybe<Array<Maybe<StaffStats>>>;
  favouredStudios?: Maybe<Array<Maybe<StudioStats>>>;
  favouredTags?: Maybe<Array<Maybe<TagStats>>>;
  favouredYears?: Maybe<Array<Maybe<YearStats>>>;
  mangaListScores?: Maybe<ListScoreStats>;
  mangaScoreDistribution?: Maybe<Array<Maybe<ScoreDistribution>>>;
  mangaStatusDistribution?: Maybe<Array<Maybe<StatusDistribution>>>;
  /** The amount of anime the user has watched in minutes */
  watchedTime?: Maybe<Scalars["Int"]>;
};